relocate merge_ordered_dicts
